refactor: organize internal packages
