+++ v1/internal/game/buildings/barracks.go
