+++ v1/internal/game/buildings/barracks_test.go
