+++ v1/internal/game/buildings/base_test.go
