+++ v1/internal/game/buildings/miner.go
