+++ v1/internal/game/buildings/miner_test.go
