+++ v1/internal/game/core/entity.go
