+++ v1/internal/game/core/entity_test.go
