+++ v1/internal/game/core/point.go
