+++ v1/internal/game/core/point_test.go
