+++ v1/internal/game/core/timer_test.go
