+++ v1/internal/game/units/footman_test.go
